RBP: Register ProcessInfo
