+++ cmake/treedata/rbpi/subdirs.txt
@@ -10,3 +10,4 @@ xbmc/filesystem/posix      filesystem/posix
 xbmc/utils/posix           utils_posix
 xbmc/platform/posix        posix
 xbmc/windowing/rpi         windowing/rpi # EGL
+xbmc/cores/VideoPlayer/Process/rbpi cores/VideoPlayer/Process/rbpi
