+++ xbmc/cores/VideoPlayer/Process/rbpi/CMakeLists.txt
@@ -1,0 +1,5 @@
+set(SOURCES ProcessInfoPi.cpp)
+
+set(HEADERS ProcessInfoPi.h)
+
+core_add_library(processPi)
